requirements.txt
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
httpx>=0.27.0
aiohttp==3.9.1
requests==2.31.0
